fix(spaces): address PR review feedback and separate restore logic into dedicated files
- Fix flat-meta segment manifest filename to 'flat_meta_segments.manifest.json'
- Fix KV merge restore to scan all segments via collectKVDataPaths before rewrite
- Guard RestoreAll against malformed dump files to prevent panics and return clear errors
- Enforce admin authentication requirement for dump and restore CLI commands
- Separate restore logic and unit tests into restore.go and restore_test.go
- Add unit tests for flat-meta segment rollover, KV segment merge, and malformed dump handling
