+++ internal/spaces/dump_test.go
@@ -353,122 +353,6 @@ func TestDumpSpaceFilter(t *testing.T) {
 	}
 }
 
-// TestRestoreMergeMode tests merge mode preserves existing data.
-func TestRestoreMergeMode(t *testing.T) {
-	// First create a dump from a source.
-	srcDir := t.TempDir()
-	spaceName := "merge_test"
-	spacePath := filepath.Join(srcDir, spaceName)
-	os.MkdirAll(spacePath, 0755)
-
-	meta := spaceMeta{
-		LayoutVersion: currentSpaceLayoutVersion,
-		Name:          spaceName,
-		EngineType:    "key-value",
-		IndexType:     "btree",
-	}
-	meta = normalizeSpaceMeta(meta)
-	metaBytes, _ := json.MarshalIndent(meta, "", "  ")
-	os.WriteFile(filepath.Join(spacePath, spaceMetaFileName), append(metaBytes, '\n'), 0644)
-
-	srcData, _ := os.Create(filepath.Join(spacePath, "data.db"))
-	writeKVRecord(t, srcData, "shared_key", "dump_value")
-	writeKVRecord(t, srcData, "dump_only", "from_dump")
-	srcData.Close()
-
-	var dumpBuf bytes.Buffer
-	_, err := DumpAll(srcDir, &dumpBuf, "")
-	if err != nil {
-		t.Fatalf("DumpAll failed: %v", err)
-	}
-
-	// Create destination with existing data.
-	dstDir := t.TempDir()
-	dstPath := filepath.Join(dstDir, spaceName)
-	os.MkdirAll(dstPath, 0755)
-	os.WriteFile(filepath.Join(dstPath, spaceMetaFileName), append(metaBytes, '\n'), 0644)
-
-	dstData, _ := os.Create(filepath.Join(dstPath, "data.db"))
-	writeKVRecord(t, dstData, "shared_key", "existing_value")
-	writeKVRecord(t, dstData, "existing_only", "from_existing")
-	dstData.Close()
-
-	// Restore in merge mode.
-	stats, err := RestoreAll(dstDir, bytes.NewReader(dumpBuf.Bytes()), "merge")
-	if err != nil {
-		t.Fatalf("RestoreAll merge failed: %v", err)
-	}
-	if stats.KVRecords != 3 {
-		t.Errorf("expected 3 merged KV records, got %d", stats.KVRecords)
-	}
-
-	// Verify merged data.
-	restored := make(map[string]string)
-	restoredDel := make(map[string]bool)
-	scanKVDataFile(filepath.Join(dstDir, spaceName, "data.db"), restored, restoredDel)
-
-	if restored["shared_key"] != "dump_value" {
-		t.Errorf("shared_key: expected 'dump_value' (dump wins), got %q", restored["shared_key"])
-	}
-	if restored["dump_only"] != "from_dump" {
-		t.Errorf("dump_only: expected 'from_dump', got %q", restored["dump_only"])
-	}
-	if restored["existing_only"] != "from_existing" {
-		t.Errorf("existing_only: expected 'from_existing', got %q", restored["existing_only"])
-	}
-}
-
-// TestRestoreInvalidVersion tests that restore rejects incompatible dump versions.
-func TestRestoreInvalidVersion(t *testing.T) {
-	header := DumpHeader{Version: 999, CreatedAt: "2026-01-01T00:00:00Z"}
-	data, _ := json.Marshal(header)
-	data = append(data, '\n')
-
-	_, err := RestoreAll(t.TempDir(), bytes.NewReader(data), "overwrite")
-	if err == nil {
-		t.Error("expected error for unsupported dump version")
-	}
-	if !strings.Contains(err.Error(), "unsupported dump format version") {
-		t.Errorf("unexpected error: %v", err)
-	}
-}
-
-// TestRestoreInvalidMode tests that restore rejects invalid mode values.
-func TestRestoreInvalidMode(t *testing.T) {
-	header := DumpHeader{Version: DumpFormatVersion, CreatedAt: "2026-01-01T00:00:00Z"}
-	data, _ := json.Marshal(header)
-	data = append(data, '\n')
-
-	_, err := RestoreAll(t.TempDir(), bytes.NewReader(data), "invalid")
-	if err == nil {
-		t.Error("expected error for invalid mode")
-	}
-}
-
-// TestUpdateManifestAfterRestore tests manifest rewrite.
-func TestUpdateManifestAfterRestore(t *testing.T) {
-	baseDir := t.TempDir()
-
-	// Create two space dirs with metadata.
-	for _, name := range []string{"alpha", "beta"} {
-		spacePath := filepath.Join(baseDir, name)
-		os.MkdirAll(spacePath, 0755)
-		meta := spaceMeta{Name: name, EngineType: "key-value", IndexType: "btree"}
-		metaBytes, _ := json.MarshalIndent(normalizeSpaceMeta(meta), "", "  ")
-		os.WriteFile(filepath.Join(spacePath, spaceMetaFileName), append(metaBytes, '\n'), 0644)
-	}
-
-	if err := UpdateManifestAfterRestore(baseDir); err != nil {
-		t.Fatalf("UpdateManifestAfterRestore failed: %v", err)
-	}
-
-	// Verify manifest.
-	manifestPath := filepath.Join(baseDir, spacesManifestFileName)
-	if _, err := os.Stat(manifestPath); err != nil {
-		t.Fatalf("manifest missing: %v", err)
-	}
-}
-
 // --- Test helpers ---
 
 func writeKVRecord(t *testing.T, f *os.File, key, value string) {
